refactor(driver): extract AArch64 window search inputs into a pure tested seam
Move the three live AArch64 per-window search-input functions —
validate_basic_block, aarch64_search_registers, and
live_out_for_optimization_prefix — out of the main.rs binary crate into a new
s11::aarch64_search_inputs library module, mirroring the x86 sibling
s11::x86_search_inputs (#752).

These functions carry the soundness-critical liveness/narrowing and
single-basic-block (issue #69) rules but lived in the binary, reachable only
through inline unit tests. Lifting them behind a pure seam (library types in,
library types out; no CLI, ELF, or Capstone) makes each rule reachable from
integration tests and keeps the driver a thin adapter. The move is
behaviour-preserving: the functions and their names are unchanged and their 12
unit tests move with them. A new tests/integration/aarch64_search_inputs.rs
proves the seam is reachable from outside the crate.

Deferred follow-ups (excluded to keep a behaviour-identical move): rename to
match the x86 sibling (validate_basic_block -> validate_terminator_placement,
aarch64_search_registers -> registers_from_target,
live_out_for_optimization_prefix -> live_out_for_optimization); and make the
AArch64 enumerative immediates target-derived like the x86 path.
