+++ tests/integration/mod.rs
@@ -1,3 +1,4 @@
+mod aarch64_search_inputs;
 mod bump_version_test;
 mod carry_chain;
 mod disasm_test;
